requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
elasticsearch < 8.0.0
opentelemetry-exporter-otlp